fixing cursor bug

+++ proj/src/model/sprite.h
@@ -17,8 +17,8 @@
 typedef struct {
     uint16_t height;
     uint16_t width;
-    uint16_t x;
-    uint16_t y;
+    int x;
+    int y;
     uint32_t *colors;
     uint32_t color;
 } Sprite;
